+++ src/plugins/deleteDecorator/calculate/regex.ts
